Allocates and frees physical memory.

+++ src/lib.hpp
@@ -9,7 +9,7 @@ namespace Runtime
 extern "C"
 {
 	void memcpy(void *dst, const void *src, size_t size);
-	void *memset(void *ptr, int value, unsigned long num);
+	void *memset(void *ptr, uint8_t value, unsigned long num);
 	unsigned long strcmp(const char *str1, const char *str2);
 	unsigned long strlen(const char *src);
 	bool strrcmp(const char *start, const char *stop, const char *str);
